Rename `extractor` module to `checker`
