+++ checker/src/test/scala/io/sigs/seals/checker/testTypes.scala
@@ -15,7 +15,7 @@
  */
 
 package io.sigs.seals
-package extractor
+package checker
 
 @core.schema
 sealed trait Foo
